feat(runtime-vapor): extract apiSetup & implement setupContext
