+++ packages/runtime-vapor/src/apiRender.ts
@@ -1,15 +1,8 @@
-import { isArray, isFunction, isObject } from '@vue/shared'
-import {
-  type ComponentInternalInstance,
-  componentKey,
-  setCurrentInstance,
-} from './component'
+import type { ComponentInternalInstance } from './component'
 import { insert, querySelector, remove } from './dom/element'
 import { flushPostFlushCbs, queuePostRenderEffect } from './scheduler'
-import { proxyRefs } from '@vue/reactivity'
 import { invokeLifecycle } from './componentLifecycle'
 import { VaporLifecycleHooks } from './apiLifecycle'
-import { fallThroughAttrs } from './componentAttrs'
 
 export const fragmentKey = Symbol(__DEV__ ? `fragmentKey` : ``)
 
@@ -18,49 +11,6 @@ export type Fragment = {
   nodes: Block
   anchor?: Node
   [fragmentKey]: true
-}
-
-export function setupComponent(
-  instance: ComponentInternalInstance,
-  singleRoot: boolean = false,
-): void {
-  const reset = setCurrentInstance(instance)
-  instance.scope.run(() => {
-    const { component, props, emit, attrs } = instance
-    const ctx = { expose: () => {}, emit, attrs }
-
-    const setupFn = isFunction(component) ? component : component.setup
-    const stateOrNode = setupFn && setupFn(props, ctx)
-
-    let block: Block | undefined
-
-    if (
-      stateOrNode &&
-      (stateOrNode instanceof Node ||
-        isArray(stateOrNode) ||
-        fragmentKey in stateOrNode ||
-        componentKey in stateOrNode)
-    ) {
-      block = stateOrNode
-    } else if (isObject(stateOrNode)) {
-      instance.setupState = proxyRefs(stateOrNode)
-    }
-    if (!block && component.render) {
-      block = component.render(instance.setupState)
-    }
-
-    if (block instanceof DocumentFragment) {
-      block = Array.from(block.childNodes)
-    }
-    if (!block) {
-      // TODO: warn no template
-      block = []
-    }
-    instance.block = block
-    if (singleRoot) fallThroughAttrs(instance)
-    return block
-  })
-  reset()
 }
 
 export function render(
